editor: fix last remaining compiler warnings

+++ source/window.c
@@ -108,7 +108,7 @@ void keyPress(GLFWwindow *window, int key, int scancode, int action, int mods)
       /* note: this way of doing things has to go away, since
        * we should use the forthcoming Text api to scroll */
       editor->cursorCol += 1;
-      if (editor->cursorCol >= editor->lineBytelen)
-         editor->cursorCol = editor->lineBytelen - 1;
+      if ((u32) editor->cursorCol >= editor->lineBytelen)
+         editor->cursorCol = (i32) editor->lineBytelen - 1;
    }
 }
